+++ common/src/main/java/com/agricraft/agricraft/common/item/CropSticksItem.java
@@ -2,7 +2,9 @@ package com.agricraft.agricraft.common.item;
 
 import com.agricraft.agricraft.common.block.CropBlock;
 import com.agricraft.agricraft.common.block.CropStickVariant;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
 import net.minecraft.tags.DamageTypeTags;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
@@ -11,10 +13,13 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+
+import java.util.List;
 
 public class CropSticksItem extends BlockItem {
 
@@ -62,6 +67,16 @@ public class CropSticksItem extends BlockItem {
 	@Override
 	public String getDescriptionId() {
 		return "item.agricraft." + variant.getSerializedName() + "_crop_sticks";
+	}
+
+	@Override
+	public void appendHoverText(ItemStack stack, Item.TooltipContext context, List<Component> tooltipComponents, TooltipFlag isAdvanced) {
+		super.appendHoverText(stack, context, tooltipComponents, isAdvanced);
+		switch (this.variant) {
+			case IRON -> tooltipComponents.add(Component.translatable("agricraft.tooltip.crop_sticks.weed_resistance", 30).withStyle(ChatFormatting.DARK_GRAY));
+			case OBSIDIAN -> tooltipComponents.add(Component.translatable("agricraft.tooltip.crop_sticks.weed_resistance", 60).withStyle(ChatFormatting.DARK_GRAY));
+			case WOODEN -> { /* no bonus */ }
+		}
 	}
 
 }